[Tutorial] implement tutorial header
