+++ .prettierrc
@@ -1,0 +1,28 @@
+{
+	"useTabs": false,
+	"indentSize": 2,
+	"indentStyle": "space",
+	"singleQuote": true,
+	"trailingComma": "es5",
+	"printWidth": 80,
+	"overrides": [
+		{
+			"files": "*.json",
+			"options": {
+				"parser": "json"
+			}
+		},
+		{
+			"files": "*.js",
+			"options": {
+				"parser": "babel"
+			}
+		},
+		{
+			"files": "*.css",
+			"options": {
+				"parser": "css"
+			}
+		}
+	]
+}
