+++ components/Tutorial/index.js
@@ -1,0 +1,1 @@
+export { default as TutorialHeader } from './TutorialHeader';
